+++ src/components/About.vue
@@ -2,7 +2,7 @@
   <div class="info-container">
     <br />
     <div>Contact: <a href="mailto:daerwendeshu@outlook.com">daerwendeshu@outlook.com</a></div>
-    <div><a href="https://gist.github.com/darwintree/284f847789a3d881d53245aee30bf259/raw/bf68ee9c261244515aacb3836319d688cc293b51/viewer-button.user.js">Github Translate Viewer Link Extension</a></div>
+    <div><a href="./viewer-button.user.js">Github Translate Viewer Link Extension</a></div>
     <div>Special Thanks: <a href=" https://github.com/ShinyColorsDB/ShinyColorsDB-EventViewer">ShinyColorsDB-EventViewer</a></div>
   </div>
 </template>